refactor: user와 관련된 요청을 user 요청만 있는 파일로 이동
