+++ client/src/apis/request/user.ts
@@ -1,20 +1,32 @@
-import {User} from 'types/serviceType';
+import {User, UserInfo} from 'types/serviceType';
 
 import {BASE_URL} from '@apis/baseUrl';
 import {USER_API_PREFIX} from '@apis/endpointPrefix';
-import {requestDelete} from '@apis/fetcher';
-import {requestGet} from '@apis/fetcher';
+import {requestDelete, requestPatchWithoutResponse} from '@apis/request';
+import {requestGet} from '@apis/request';
 
 export const requestDeleteUser = async () => {
   await requestDelete({
     baseUrl: BASE_URL.HD,
-    endpoint: `${USER_API_PREFIX}`,
+    endpoint: USER_API_PREFIX,
   });
 };
 
 export const requestGetUserInfo = async () => {
-  return await requestGet<User>({
+  return await requestGet<UserInfo>({
     baseUrl: BASE_URL.HD,
-    endpoint: `/api/users/mine`,
+    endpoint: `${USER_API_PREFIX}/mine`,
+    errorHandlingStrategy: 'unsubscribe',
+  });
+};
+
+export type RequestPatchUser = Partial<User>;
+
+export const requestPatchUser = async (args: RequestPatchUser) => {
+  requestPatchWithoutResponse({
+    endpoint: USER_API_PREFIX,
+    body: {
+      ...args,
+    },
   });
 };
